shekel-watch-python/requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
yfinance>=0.2.36
pandas>=2.0.0
plotly>=5.18.0
crewai>=0.80.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain-openai>=0.2.0
